vision_with_depth_v15

+++ look_to_point/src/can_dtct.cpp
@@ -122,7 +122,7 @@ ros::Time latestImageStamp;
 // My Function of detecting the top of cans
 
 
-int ReadDepthData(unsigned int height_pos, unsigned int width_pos, sensor_msgs::ImageConstPtr depth_image)
+double ReadDepthData(unsigned int height_pos, unsigned int width_pos, sensor_msgs::ImageConstPtr depth_image)
 {
     // If position is invalid
     if ((height_pos >= depth_image->height) || (width_pos >= depth_image->width))
@@ -140,7 +140,7 @@ int ReadDepthData(unsigned int height_pos, unsigned int width_pos, sensor_msgs::
               depth_data.byte_data[i] = depth_image->data[index + i];
 
           if (depth_data.float_data == depth_data.float_data)
-              return int(depth_data.float_data);
+              return double(depth_data.float_data);
 
           return -1;  // If depth data invalid
         }
@@ -150,7 +150,7 @@ int ReadDepthData(unsigned int height_pos, unsigned int width_pos, sensor_msgs::
             depth_data.byte_data[i] = depth_image->data[3 + index - i];
         // Make sure data is valid (check if NaN)
         if (depth_data.float_data == depth_data.float_data)
-            return int(depth_data.float_data);
+            return double(depth_data.float_data);
         return -1;  // If depth data invalid
     }
     // Otherwise, data is 2 byte integers (raw depth image)
